+++ platform/android/app/src/main/cpp/quest_android_shims.c
@@ -33,6 +33,7 @@ uint8_t gRenderingInterpolated;
 static char sQuestUserPath[4096] =
     "/sdcard/Android/data/com.fulldivegames.sm64coopdxvr/files";
 static const char sQuestSharedModPath[] = "/sdcard/SM64VR/mods";
+static const char sQuestSharedDynosPackPath[] = "/sdcard/SM64VR/dynos/packs";
 char gRomFilename[4096] =
     "/sdcard/Android/data/com.fulldivegames.sm64coopdxvr/files/baserom.us.z64";
 bool gGameInited;
@@ -176,6 +177,7 @@ void quest_android_set_user_path(const char *path) {
 
 const char *sys_user_path(void) { return sQuestUserPath; }
 const char *quest_android_shared_mod_path(void) { return sQuestSharedModPath; }
+const char *quest_android_shared_dynos_pack_path(void) { return sQuestSharedDynosPackPath; }
 const char *sys_resource_path(void) { return sys_user_path(); }
 const char *sys_exe_path_dir(void) { return sys_user_path(); }
 const char *sys_exe_path_file(void) { return "/proc/self/exe"; }